+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.7.1-SNAPSHOT]
 
-
+- TextInputControl.trimWhitespace() enforces on focus lost instead of onChange (prevented adding words with whitespace)
 
 ## [1.7.0] - 2017-03-04